完成模组领域对象

+++ northstar-strategy-common/src/main/java/tech/xuanwu/northstar/strategy/common/model/ModuleInfo.java
@@ -36,9 +36,9 @@ public class ModuleInfo {
 	@Id
 	private String moduleName;
 	/**
-	 * 账户比例
+	 * 账户比例(单位:%)
 	 */
-	private double allocatedAccountShare;
+	private int allocatedAccountShare;
 	/**
 	 * 是否启用
 	 */
